requirements.txt
transformers
torch
pandas
tqdm
openai
google-generativeai
scikit-learn
